- 2022/11/15

+++ ver_2/include/TrajectoryGenerator.h
@@ -5,5 +5,5 @@
 using namespace std;
 
 const int MILLIS_TO_NEXT_FRAME = 20;
-vector<vector<float>> GenParaBlendForCableMotor(double start[6], double end[6], int time, bool showAttention);
+vector<vector<double>> GenParaBlendForCableMotor(double start[6], double end[6], int time, bool showAttention);
 #endif
